extract boolean expression stream logic and reuse it to sample with in collector
