+++ lib/page_creator/assets/common.dart
@@ -15,29 +15,6 @@ part 'common.g.dart';
 
 const String constAssetName = "asset_name";
 
-@JsonSerializable()
-class ColorConverter implements JsonConverter<Color, Map<String, dynamic>> {
-  const ColorConverter();
-
-  @override
-  Color fromJson(Map<String, dynamic> json) {
-    return Color.fromRGBO(
-      (json['red']! * 255).toInt(),
-      (json['green']! * 255).toInt(),
-      (json['blue']! * 255).toInt(),
-      json['alpha'] ?? 1.0,
-    );
-  }
-
-  @override
-  Map<String, double> toJson(Color color) => {
-        'red': color.r,
-        'green': color.g,
-        'blue': color.b,
-        'alpha': color.a,
-      };
-}
-
 @JsonEnum()
 enum TextPos {
   above,
