+++ lib/page_creator/assets/led.dart
@@ -9,6 +9,7 @@ import 'package:rxdart/rxdart.dart';
 
 import 'common.dart';
 import '../../providers/state_man.dart';
+import '../../converter/color_converter.dart';
 
 part 'led.g.dart';
