+++ test/core/duration_converter_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:tfc/core/duration_converter.dart';
+import 'package:tfc/converter/duration_converter.dart';
 
 void main() {
   group('DurationMicrosecondsConverter', () {
